mrc-leaf-helpers(step-span-name-move,mechanical_provable): Move build_step_span_name to utils.profile_utils (cut+paste)
